backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pymysql==1.1.1
cryptography==42.0.7
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
pytest==8.2.0
httpx==0.27.0
